refactor: decompose discord/bridge.ts from 2688 to 367 lines (#932)
Extract 6 focused modules from the monolithic bridge.ts:
- commands.ts: slash command registration & interaction handling
- admin-commands.ts: /admin subcommand handlers
- embeds.ts: Discord API helpers, embed builders, messaging utilities
- message-handler.ts: message routing, work intake, mention replies
- permissions.ts: RBAC, rate limiting, channel monitoring
- thread-manager.ts: thread lifecycle, response streaming, archival

Bridge.ts retained as thin orchestration layer owning state and
delegating to extracted modules via context objects.

No behavioral changes — pure structural refactoring.
