+++ server/discord/admin-commands.ts
@@ -1,0 +1,392 @@
+/**
+ * Discord admin command handlers.
+ *
+ * Handles `/admin channels|users|roles|mode|public|show` subcommands.
+ * All changes persist to `discord_config` DB table and hot-reload within 30s.
+ */
+
+import type { Database } from 'bun:sqlite';
+import type { DiscordBridgeConfig, DiscordInteractionData, DiscordInteractionOption } from './types';
+import { respondToInteraction, respondToInteractionEmbed } from './embeds';
+import { updateDiscordConfig } from '../db/discord-config';
+import { recordAudit } from '../db/audit';
+
+/** Discord snowflake IDs are purely numeric strings. */
+const DISCORD_SNOWFLAKE_RE = /^\d{17,20}$/;
+
+function levelName(level: number): string {
+    return level === 0 ? 'Blocked' : level === 1 ? 'Basic' : level === 2 ? 'Standard' : level === 3 ? 'Admin' : `Level ${level}`;
+}
+
+export async function handleAdminCommand(
+    db: Database,
+    config: DiscordBridgeConfig,
+    mutedUsers: Set<string>,
+    threadSessionCount: number,
+    interaction: DiscordInteractionData,
+    options: DiscordInteractionOption[],
+): Promise<void> {
+    const group = options[0];
+    if (!group) {
+        await respondToInteraction(interaction, 'Missing subcommand.');
+        return;
+    }
+
+    const groupName = group.name;
+
+    // Direct subcommands (mode, public, show)
+    if (groupName === 'show') {
+        await handleAdminShow(config, mutedUsers, threadSessionCount, interaction);
+        return;
+    }
+
+    if (groupName === 'mode') {
+        const sub = group.options?.[0];
+        const mode = sub?.value as string ?? group.value as string;
+        if (mode !== 'chat' && mode !== 'work_intake') {
+            await respondToInteraction(interaction, 'Invalid mode. Use `chat` or `work_intake`.');
+            return;
+        }
+        updateDiscordConfig(db, 'mode', mode);
+        config.mode = mode;
+        recordAudit(db, 'discord_config_update', interaction.member?.user?.id ?? 'unknown', 'discord_config', 'mode', JSON.stringify({ value: mode }));
+        await respondToInteractionEmbed(interaction, {
+            title: 'Bridge Mode Updated',
+            description: `Mode set to **${mode === 'chat' ? 'Chat' : 'Work Intake'}**`,
+            color: 0x57f287,
+            footer: { text: mode === 'chat' ? 'Messages route to agent sessions' : 'Messages create async work tasks' },
+        });
+        return;
+    }
+
+    if (groupName === 'public') {
+        const sub = group.options?.[0];
+        const enabled = (sub?.value ?? group.value) as boolean;
+        updateDiscordConfig(db, 'public_mode', String(enabled));
+        config.publicMode = enabled;
+        recordAudit(db, 'discord_config_update', interaction.member?.user?.id ?? 'unknown', 'discord_config', 'public_mode', JSON.stringify({ value: enabled }));
+        await respondToInteractionEmbed(interaction, {
+            title: 'Public Mode Updated',
+            description: enabled
+                ? 'Public mode **enabled** — all users can interact (subject to role permissions)'
+                : 'Public mode **disabled** — only allowed users can interact',
+            color: enabled ? 0x57f287 : 0xed4245,
+        });
+        return;
+    }
+
+    // Subcommand groups (channels, users, roles)
+    const subcommand = group.options?.[0];
+    if (!subcommand) {
+        await respondToInteraction(interaction, 'Missing subcommand.');
+        return;
+    }
+    const subName = subcommand.name;
+    const subOptions = subcommand.options ?? [];
+    const getSubOption = (name: string) => subOptions.find(o => o.name === name)?.value;
+
+    switch (groupName) {
+        case 'channels':
+            await handleAdminChannels(db, config, interaction, subName, getSubOption);
+            break;
+        case 'users':
+            await handleAdminUsers(db, config, interaction, subName, getSubOption);
+            break;
+        case 'roles':
+            await handleAdminRoles(db, config, interaction, subName, getSubOption);
+            break;
+        default:
+            await respondToInteraction(interaction, `Unknown admin subcommand: ${groupName}`);
+    }
+}
+
+async function handleAdminChannels(
+    db: Database,
+    config: DiscordBridgeConfig,
+    interaction: DiscordInteractionData,
+    subName: string,
+    getSubOption: (name: string) => string | number | boolean | undefined,
+): Promise<void> {
+    const current = config.additionalChannelIds ?? [];
+
+    switch (subName) {
+        case 'add': {
+            const channelId = String(getSubOption('channel'));
+            if (!channelId || !DISCORD_SNOWFLAKE_RE.test(channelId)) {
+                await respondToInteraction(interaction, 'Invalid channel.');
+                return;
+            }
+            if (channelId === config.channelId) {
+                await respondToInteraction(interaction, 'That is already the primary channel.');
+                return;
+            }
+            if (current.includes(channelId)) {
+                await respondToInteraction(interaction, `<#${channelId}> is already monitored.`);
+                return;
+            }
+            const updated = [...current, channelId];
+            updateDiscordConfig(db, 'additional_channel_ids', updated.join(','));
+            config.additionalChannelIds = updated;
+            recordAudit(db, 'discord_config_update', interaction.member?.user?.id ?? 'unknown', 'discord_config', 'additional_channel_ids', JSON.stringify({ action: 'add', channelId }));
+            await respondToInteractionEmbed(interaction, {
+                title: 'Channel Added',
+                description: `<#${channelId}> is now being monitored.\n\nTotal monitored: **${updated.length + 1}** (including primary)`,
+                color: 0x57f287,
+            });
+            break;
+        }
+        case 'remove': {
+            const channelId = String(getSubOption('channel'));
+            if (!channelId || !DISCORD_SNOWFLAKE_RE.test(channelId)) {
+                await respondToInteraction(interaction, 'Invalid channel.');
+                return;
+            }
+            if (channelId === config.channelId) {
+                await respondToInteraction(interaction, 'Cannot remove the primary channel. Change `DISCORD_CHANNEL_ID` in your environment to change it.');
+                return;
+            }
+            if (!current.includes(channelId)) {
+                await respondToInteraction(interaction, `<#${channelId}> is not in the monitored list.`);
+                return;
+            }
+            const updated = current.filter(id => id !== channelId);
+            updateDiscordConfig(db, 'additional_channel_ids', updated.join(','));
+            config.additionalChannelIds = updated;
+            recordAudit(db, 'discord_config_update', interaction.member?.user?.id ?? 'unknown', 'discord_config', 'additional_channel_ids', JSON.stringify({ action: 'remove', channelId }));
+            await respondToInteractionEmbed(interaction, {
+                title: 'Channel Removed',
+                description: `<#${channelId}> removed from monitoring.\n\nTotal monitored: **${updated.length + 1}** (including primary)`,
+                color: 0xed4245,
+            });
+            break;
+        }
+        case 'list': {
+            const allChannels = [
+                `<#${config.channelId}> *(primary)*`,
+                ...current.map(id => `<#${id}>`),
+            ];
+            await respondToInteractionEmbed(interaction, {
+                title: 'Monitored Channels',
+                description: allChannels.join('\n') || 'No channels configured.',
+                color: 0x5865f2,
+                footer: { text: `${allChannels.length} channel${allChannels.length === 1 ? '' : 's'} total` },
+            });
+            break;
+        }
+    }
+}
+
+async function handleAdminUsers(
+    db: Database,
+    config: DiscordBridgeConfig,
+    interaction: DiscordInteractionData,
+    subName: string,
+    getSubOption: (name: string) => string | number | boolean | undefined,
+): Promise<void> {
+    const current = [...config.allowedUserIds];
+
+    switch (subName) {
+        case 'add': {
+            const userId = String(getSubOption('user'));
+            if (!userId || !DISCORD_SNOWFLAKE_RE.test(userId)) {
+                await respondToInteraction(interaction, 'Invalid user.');
+                return;
+            }
+            if (current.includes(userId)) {
+                await respondToInteraction(interaction, `<@${userId}> is already on the allow list.`);
+                return;
+            }
+            const updated = [...current, userId];
+            updateDiscordConfig(db, 'allowed_user_ids', updated.join(','));
+            config.allowedUserIds = updated;
+            recordAudit(db, 'discord_config_update', interaction.member?.user?.id ?? 'unknown', 'discord_config', 'allowed_user_ids', JSON.stringify({ action: 'add', userId }));
+            await respondToInteractionEmbed(interaction, {
+                title: 'User Added',
+                description: `<@${userId}> added to the allow list.\n\nTotal allowed: **${updated.length}**`,
+                color: 0x57f287,
+            });
+            break;
+        }
+        case 'remove': {
+            const userId = String(getSubOption('user'));
+            if (!userId || !DISCORD_SNOWFLAKE_RE.test(userId)) {
+                await respondToInteraction(interaction, 'Invalid user.');
+                return;
+            }
+            if (!current.includes(userId)) {
+                await respondToInteraction(interaction, `<@${userId}> is not on the allow list.`);
+                return;
+            }
+            const updated = current.filter(id => id !== userId);
+            updateDiscordConfig(db, 'allowed_user_ids', updated.join(','));
+            config.allowedUserIds = updated;
+            recordAudit(db, 'discord_config_update', interaction.member?.user?.id ?? 'unknown', 'discord_config', 'allowed_user_ids', JSON.stringify({ action: 'remove', userId }));
+            await respondToInteractionEmbed(interaction, {
+                title: 'User Removed',
+                description: `<@${userId}> removed from the allow list.\n\nTotal allowed: **${updated.length}**`,
+                color: 0xed4245,
+            });
+            break;
+        }
+        case 'list': {
+            const userLines = current.length > 0
+                ? current.map(id => `<@${id}>`).join('\n')
+                : '_No users on allow list — all users have access (legacy mode with empty list)_';
+            await respondToInteractionEmbed(interaction, {
+                title: 'Allowed Users',
+                description: userLines,
+                color: 0x5865f2,
+                footer: { text: `${current.length} user${current.length === 1 ? '' : 's'} · ${config.publicMode ? 'Public mode (roles take precedence)' : 'Legacy mode (allow list enforced)'}` },
+            });
+            break;
+        }
+    }
+}
+
+async function handleAdminRoles(
+    db: Database,
+    config: DiscordBridgeConfig,
+    interaction: DiscordInteractionData,
+    subName: string,
+    getSubOption: (name: string) => string | number | boolean | undefined,
+): Promise<void> {
+    const current = { ...(config.rolePermissions ?? {}) };
+
+    switch (subName) {
+        case 'set': {
+            const roleId = String(getSubOption('role'));
+            const level = Number(getSubOption('level'));
+            if (!roleId || !DISCORD_SNOWFLAKE_RE.test(roleId)) {
+                await respondToInteraction(interaction, 'Invalid role.');
+                return;
+            }
+            if (level < 0 || level > 3 || !Number.isInteger(level)) {
+                await respondToInteraction(interaction, 'Permission level must be 0-3.');
+                return;
+            }
+            current[roleId] = level;
+            const json = JSON.stringify(current);
+            updateDiscordConfig(db, 'role_permissions', json);
+            config.rolePermissions = current;
+            recordAudit(db, 'discord_config_update', interaction.member?.user?.id ?? 'unknown', 'discord_config', 'role_permissions', JSON.stringify({ action: 'set', roleId, level }));
+            await respondToInteractionEmbed(interaction, {
+                title: 'Role Permission Set',
+                description: `<@&${roleId}> → **${levelName(level)}** (${level})`,
+                color: 0x57f287,
+                footer: { text: `${Object.keys(current).length} role mapping${Object.keys(current).length === 1 ? '' : 's'} configured` },
+            });
+            break;
+        }
+        case 'remove': {
+            const roleId = String(getSubOption('role'));
+            if (!roleId || !DISCORD_SNOWFLAKE_RE.test(roleId)) {
+                await respondToInteraction(interaction, 'Invalid role.');
+                return;
+            }
+            if (!(roleId in current)) {
+                await respondToInteraction(interaction, `<@&${roleId}> has no permission override.`);
+                return;
+            }
+            delete current[roleId];
+            const json = JSON.stringify(current);
+            updateDiscordConfig(db, 'role_permissions', json);
+            config.rolePermissions = current;
+            recordAudit(db, 'discord_config_update', interaction.member?.user?.id ?? 'unknown', 'discord_config', 'role_permissions', JSON.stringify({ action: 'remove', roleId }));
+            await respondToInteractionEmbed(interaction, {
+                title: 'Role Permission Removed',
+                description: `<@&${roleId}> permission override removed.\nUsers with this role will get the default level (**${levelName(config.defaultPermissionLevel ?? 1)}**).`,
+                color: 0xed4245,
+            });
+            break;
+        }
+        case 'list': {
+            const entries = Object.entries(current);
+            const roleLines = entries.length > 0
+                ? entries.map(([rid, lvl]) => `<@&${rid}> → **${levelName(lvl)}** (${lvl})`).join('\n')
+                : '_No role permissions configured_';
+            await respondToInteractionEmbed(interaction, {
+                title: 'Role Permissions',
+                description: roleLines,
+                color: 0x5865f2,
+                fields: [
+                    {
+                        name: 'Default Level',
+                        value: `**${levelName(config.defaultPermissionLevel ?? 1)}** (${config.defaultPermissionLevel ?? 1})`,
+                        inline: true,
+                    },
+                    {
+                        name: 'Public Mode',
+                        value: config.publicMode ? 'Enabled' : 'Disabled',
+                        inline: true,
+                    },
+                ],
+                footer: { text: entries.length > 0 ? `${entries.length} role mapping${entries.length === 1 ? '' : 's'}` : 'Use /admin roles set to add mappings' },
+            });
+            break;
+        }
+    }
+}
+
+async function handleAdminShow(
+    config: DiscordBridgeConfig,
+    mutedUsers: Set<string>,
+    threadSessionCount: number,
+    interaction: DiscordInteractionData,
+): Promise<void> {
+    const channels = [
+        `<#${config.channelId}> *(primary)*`,
+        ...(config.additionalChannelIds ?? []).map(id => `<#${id}>`),
+    ];
+    const users = config.allowedUserIds.length > 0
+        ? config.allowedUserIds.map(id => `<@${id}>`).join(', ')
+        : '_none_';
+    const roleEntries = Object.entries(config.rolePermissions ?? {});
+    const roles = roleEntries.length > 0
+        ? roleEntries.map(([rid, lvl]) => `<@&${rid}>=${levelName(lvl)}`).join(', ')
+        : '_none_';
+
+    await respondToInteractionEmbed(interaction, {
+        title: 'Bot Configuration',
+        color: 0x5865f2,
+        fields: [
+            {
+                name: 'Mode',
+                value: config.mode === 'work_intake' ? 'Work Intake' : 'Chat',
+                inline: true,
+            },
+            {
+                name: 'Public Mode',
+                value: config.publicMode ? 'Enabled' : 'Disabled',
+                inline: true,
+            },
+            {
+                name: 'Default Permission',
+                value: `${levelName(config.defaultPermissionLevel ?? 1)} (${config.defaultPermissionLevel ?? 1})`,
+                inline: true,
+            },
+            {
+                name: `Channels (${channels.length})`,
+                value: channels.join('\n').slice(0, 1024),
+                inline: false,
+            },
+            {
+                name: 'Allowed Users',
+                value: users.slice(0, 1024),
+                inline: false,
+            },
+            {
+                name: 'Role Permissions',
+                value: roles.slice(0, 1024),
+                inline: false,
+            },
+            {
+                name: 'Muted Users',
+                value: mutedUsers.size > 0
+                    ? [...mutedUsers].map(id => `<@${id}>`).join(', ').slice(0, 1024)
+                    : '_none_',
+                inline: false,
+            },
+        ],
+        footer: { text: `Active sessions: ${threadSessionCount} · Config reloads every 30s` },
+    });
+}
